requirements.txt
pytest
pytest-cov
python-coveralls

flask
flask_sqlalchemy
flask-login
flask-restless-continued
pyyaml
mutagen
pymediainfo
watchdog

pylast
musicbrainzngs
lyricsgenius